Add test for splitArtworkShard
