+++ .envrc
@@ -6,3 +6,11 @@ export PUBLISHER_ID="0x7bd8a9757894069470bdee4d899b34b41ed82cbd77f68ba33d0fd46d0
 export USER1_PHRASE="gossip net bamboo assist danger armed soda plastic such father better grid"
 export USER2_PHRASE="can gaze pistol praise unable man afraid include route doll evoke ticket"
 export USER3_PHRASE="inherit chat inmate swamp transfer deposit shy heart frost chest cage bonus"
+export ADMIN_CAP_ID="0xf0d43b7ec36202f000dfaf9c0c1e3ae4bfb7ae631516e3054fbfc951175bdabc"
+export ADMIN_PHRASE="future seat note injury inch feature ring kiwi soap inspire wrestle velvet"
+export PACKAGE_ID="0x06c5a856284df1dd98a9c1482607affcd85c25d3c5b1fa9cc137276bf4d6a0d2"
+export PUBLISHER_ID="0xef110567e1b2ba935480d4da6b02cf04842408905e80f28ac94dfaddbe53ecc0"
+export SUI_NETWORK="http://127.0.0.1:9000"
+export USER1_PHRASE="duck stove fruit danger erupt champion olive away music august inside worth"
+export USER2_PHRASE="enable laugh large bronze brand miracle notable shiver coyote topic ivory duck"
+export USER3_PHRASE="quick dinosaur despair cloud plastic clog thrive belt practice spoil alpha imitate"
