+++ setup/src/mintArtworkShard.ts
