+++ test/mintArtworkShard.test.ts
@@ -1,22 +1,11 @@
 import assert from "assert";
 
-import { mintArtwork, MintArtworkParams } from "../setup/src/artwork";
-import { mintArtworkShard } from "../setup/src/artwork_shard";
+import { mintArtwork } from "../setup/src/artwork";
 import { ADMIN_PHRASE, USER1_PHRASE, USER2_PHRASE } from "../setup/src/config";
+import { mintArtworkShard } from "../setup/src/mintArtworkShard";
+import { mintArtworkOptions } from "./testdata";
 
-const mintArtworkOptions: MintArtworkParams = {
-  totalSupply: 500,
-  sharePrice: 10,
-  multiplier: 100,
-  name: "Mona Lisa",
-  artist: "Leonardo da Vinci",
-  creationDate: "1685548680595",
-  description: "Choconta painting",
-  image:
-    "https://upload.wikimedia.org/wikipedia/commons/thumb/e/ec/Mona_Lisa%2C_by_Leonardo_da_Vinci%2C_from_C2RMF_retouched.jpg/800px-Mona_Lisa%2C_by_Leonardo_da_Vinci%2C_from_C2RMF_retouched.jpg",
-};
-
-describe("Artwork issue a contract", () => {
+describe("mintArtworkShard", () => {
   let artworkId: string;
   beforeEach(async () => {
     artworkId = await mintArtwork(mintArtworkOptions);
