+++ test/testdata.ts
@@ -1,0 +1,13 @@
+import { MintArtworkParams } from "../setup/src/artwork";
+
+export const mintArtworkOptions: MintArtworkParams = {
+  totalSupply: 500,
+  sharePrice: 10,
+  multiplier: 100,
+  name: "Mona Lisa",
+  artist: "Leonardo da Vinci",
+  creationDate: "1685548680595",
+  description: "Choconta painting",
+  image:
+    "https://upload.wikimedia.org/wikipedia/commons/thumb/e/ec/Mona_Lisa%2C_by_Leonardo_da_Vinci%2C_from_C2RMF_retouched.jpg/800px-Mona_Lisa%2C_by_Leonardo_da_Vinci%2C_from_C2RMF_retouched.jpg",
+};
